Adding server support.
The main SIP.js use case seems to have targeted "webrtc" and therefore
being executed within a browser. Also, it was expected that you knew
up-front the AOR etc for a given UA. However, that is not the case when
executing as a server so:

1. Added TCP as a transport.
2. Will create a UA Core based on the AOR of the incoming message.
3. Will otherwise use existing logic for dealing with transactions and
   dialogs.

Also added examples.

Finally, thank you Claude Code for doing 99% of the work!